% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binmanHelpers.R
\name{binman_path_base}
\alias{binman_path_base}
\title{Get suggested binman path}
\usage{
binman_path_base(path = here::here())
}
\arguments{
\item{path}{Path to check. Defaults to result of \code{here::here()}}
}
\value{
Returns a file path
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Get the default base path for a binman download folder. If your
working directory is in a project with a 'renv' folder, it will suggest
that as the install location. Otherwise the path specified by
\code{rappdirs::user_data_dir()} will be used.
}
\examples{
\dontrun{
binman_path_base()
}

}
