% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binmanHelpers.R
\name{binman_predl_github_assets}
\alias{binman_predl_github_assets}
\title{{binman}-like Pre download Github assets}
\usage{
binman_predl_github_assets(
  url,
  platform,
  history = 3L,
  appname,
  platformregex = platform,
  versionregex = c("", ""),
  dlPath = NULL
)
}
\arguments{
\item{url}{A url giving the github asset JSON for a project. As an example
https://github.com/mozilla/geckodriver/releases the geckodriver project has
an asset JSON available at
https://api.github.com/repos/mozilla/geckodriver/releases}

\item{platform}{A character vector of platform names}

\item{history}{The maximum number of files to get for a platform}

\item{appname}{Name of the app}

\item{platformregex}{A filter for platforms. Defaults to the platform}

\item{versionregex}{A regex for retrieving the version.}

\item{dlPath}{Custom download path for the binman folder. Defaults to NULL,
where it will return the same table as \code{binman::predl_github_assets()}
with an unmodified download path.}
}
\value{
A named list of data.frames. The name indicates the platform. The
data.frame should contain the version, url and file to be processed. Used
as input for \code{\link{download_files}} or an equivalent.
}
\description{
This is a wrapper around \code{binman::predl_github_assets()}
that adds the parameter \code{dlPath} that allows the user to specify the
binman download path. This was built to allow custom binman install
directories for project specific environments.
}
\examples{
\dontrun{
dllist <- binman_predl_github_assets(
    url = "https://api.github.com/repos/mozilla/geckodriver/releases",
    platform = c("win64"),
    history = 1L,
    appname = "geckodriver",
    versionregex = c("^v", ""),
    dlPath = here::here()
 )
dllist
}
}
