% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesceJoin.R
\name{coalesceJoin}
\alias{coalesceJoin}
\title{Coalesce Duplicate Fields from a Join}
\usage{
coalesceJoin(df, suffix = c(".x", ".y"), showMessage = TRUE)
}
\arguments{
\item{df}{the joined data.frame}

\item{suffix}{suffix used to diferentiate the columns needing coalesced. Defaults to c(".x", ".y")}

\item{showMessage}{TRUE/FALSE on whether to print message. Default TRUE.}
}
\value{
df with the join suffixed columns coalesced
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Coalesces columns by suffix created by joins. Useful after a full join.
}
\examples{
\dontrun{
#Download Case, Death, Test Data
cdtData <- MARCDataAPI_read('https://gis2.marc2.org/MARCDataAPI/api/covidcasedeathtest') \%>\%
  dplyr::mutate(Date = as.Date(Date),
                LastUpdated = lubridate::as_datetime(LastUpdated),
                LastUpdated = lubridate::with_tz(LastUpdated, "America/Chicago"))

#Download the Hospital Data
hospData <- marcR::MARCDataAPI_read('https://gis2.marc2.org/MARCDataAPI/api/covidhospital') \%>\%
    dplyr::mutate(Date = as.Date(Date))

cdtHospSumData <- dplyr::full_join(cdtData, hospData, by = c("GeoID", "Date")) \%>\% coalesceJoin()
}
}
