% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectODBC.R
\name{connectODBC}
\alias{connectODBC}
\title{Connect to ODBC Database Using keyring}
\usage{
connectODBC(databaseString = NULL)
}
\arguments{
\item{databaseString}{Same as the username saved in the keyring. Default NULL
will list the available keys under the service 'DB_conn'}
}
\value{
OBDC connection object returned from DBI::dbConnect()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Creates connection to ODBC database using the data contained in
keyrings. This is generally used for internal purposes at MARC to make
database connections a lazier proccess with {keryring}.
}
\details{
Keys should be set up prior to using the function to a style similar
to:

\code{keyring::key_set("DB_conn", username = "<serverName>.<databaseName>.<schemaName>")}

The password should then be set to the connection string formatted like:

"\verb{Driver=\{ODBC Driver 17 for SQL Server\};Server=<serverName>;Database=<databaseName>;UID=<schemaName>;PWD=<password>;}"

Leave the PWD section in the username argument written as PASSWORD as this
is automatically replaced with the stored password when calling
\code{connectODBC()}.
}
\section{Creation notes}{
 First created in 2020-Sept for easily accessing
databases in R
}

\examples{
\dontrun{
library(glue)
library(magrittr)
library(DBI)

con <- connectODBC("<serverName>.<databaseName>.<schemaName>")

DBI_getOBDCtable(con, "SELECT * FROM <schemaName>.<tableName>")
}
}
\author{
Jacob Peterson
}
