% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBIhelpers.R
\name{dbListSchema}
\alias{dbListSchema}
\title{List all schema in database}
\usage{
dbListSchema(
  conn,
  rmSchemaRegex = c("sys", "sde", "^INFORMATION_SCHEMA$", "^db_\\\\.*")
)
}
\arguments{
\item{conn}{A \code{\link[DBI:DBIConnection-class]{DBIConnection}} object, as
returned by \code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{rmSchemaRegex}{Character vector containing schema to avoid searching
(removed schema regex). Ignores some default system level schema and schema
only used by the ESRI SDE bindings that don't actually contain user created
tables.}
}
\value{
A character vector of schema's in the database connections.
}
\description{
Searches schema in in the INFORMATION_SCHEMA.SCHEMATA table.
Confirmed only to work with MS-SQL databases.
}
\examples{
\dontrun{

con <- connectODBC("<servername>.<databaseName>.<schemaName>")

#This is not set up to be a real example
schemas <- dbListSchema(con)
}
}
\author{
Jacob Peterson
}
