% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBIhelpers.R
\name{dbListTableStructure}
\alias{dbListTableStructure}
\title{List all tables in a database}
\usage{
dbListTableStructure(
  conn,
  addGeoIndicator = FALSE,
  rmTableRegex = c("^[:alpha:][:digit:]+$", "^SDE_", "^sysdiagrams$"),
  ...
)
}
\arguments{
\item{conn}{A \code{\link[DBI:DBIConnection-class]{DBIConnection}} object, as
returned by \code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{addGeoIndicator}{TRUE/FALSE. Should the \code{isSpatial} column be exported.
Default is FALSE.}

\item{rmTableRegex}{Character vector containing table names to avoid
searching (removed table regex). Ignores some tables that are only used by
the ESRI SDE bindings that don't actually contain user created data.}

\item{...}{Additional options passed to dbListSchema}
}
\value{
A dataframe with a row for each table in the database connection.
Contains 3 or 4 columns ('database', 'schema', 'table', and optionally
'isSpatial'). The return dataframe can then easily be searched, filtered,
and queried to find the tables you were looking for.
}
\description{
A more informative version of \code{DBI::dbListTables()} which only
contains table names. This functions also pairs each table with its schema
and can handle checking if the table has spatial data or not.
}
\examples{
\dontrun{

con <- connectODBC("<databaseName>.<schemaName>")

#This is not set up to be a real example
tables <- dbListTableStructure(con, addGeoIndicator = TRUE)
}
}
\author{
Jacob Peterson
}
