% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBIhelpers.R
\name{dbTableStructure}
\alias{dbTableStructure}
\title{List all tables in a database}
\usage{
dbTableStructure(
  conn,
  addGeoIndicator = FALSE,
  includeViews = TRUE,
  rmTableRegex = c("^[:alpha:][:digit:]+$", "^SDE_", "^sysdiagrams$"),
  rmSchemaRegex = c("sde")
)
}
\arguments{
\item{conn}{A \code{\link[DBI:DBIConnection-class]{DBIConnection}} object, as
returned by \code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{addGeoIndicator}{TRUE/FALSE. Should the \code{isSpatial} column be exported?
Default is FALSE.}

\item{includeViews}{TRUE/FALSE. Should output included views?}

\item{rmTableRegex}{Character vector containing table names to avoid
searching (removed table regex). Ignores some tables that are only used by
the ESRI SDE bindings that don't actually contain user created data.}

\item{rmSchemaRegex}{Character vector containing schema to avoid searching
(removed schema regex). Ignores some default system level schema and schema
only used by the ESRI SDE bindings that don't actually contain user created
tables.}
}
\value{
A dataframe with a row for each table in the database connection.
Contains 4 or 5 columns ('Database', 'Schema', 'Table', 'isView', and optionally
'isSpatial'). The return dataframe can then easily be searched, filtered,
and queried to find the tables you were looking for.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

A more informative version of \code{DBI::dbListTables()} which only
contains table names. This functions also pairs each table with its schema
and can handle checking if the table has spatial data and identify if it
is a view.
}
\examples{
\dontrun{

conn <- connectODBC("<servername>.<databaseName>.<schemaName>")

#This is not set up to be a real example
tables <- dbListTableStructure(conn, addGeoIndicator = TRUE)
}
}
\author{
Jacob Peterson
}
