% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch2Datetime.R
\name{epoch2Datetime}
\alias{epoch2Datetime}
\title{Change Epoch Time to Datetime}
\usage{
epoch2Datetime(epochDate, milliseconds = TRUE)
}
\arguments{
\item{epochDate}{numeric vector with the epoch date}

\item{milliseconds}{Should you complete millisecond conversion? Default is
TRUE}
}
\value{
Vector of class POSIXct
}
\description{
Converts epoch time to a POSIXct/datetime. This is a common task
when reading in data with dates from an ESRI REST API.
}
\section{Creation notes}{
 First created in 2020-Sept in conjunction with my
pull request to \code{esri2sf} to add the \code{esri2df} function
}

\examples{
\dontrun{
library(esri2sf)

esriData <- esri2sf::esri2df(url = "https://gis2.marc2.org/arcgis/rest/services/HumanServices/COVIDv2/MapServer/5")
class(esriData$Date)

class(epoch2Datetime(esriData$Date))
}
}
\author{
Jacob Peterson
}
