% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glueIndentLines.R
\name{glueIndentLines}
\alias{glueIndentLines}
\title{Indent lines}
\usage{
glueIndentLines(string, tabs = 1)
}
\arguments{
\item{string}{String you want to add tabs to}

\item{tabs}{Number of tabs you want to indent each line}
}
\value{
A \code{glue} character string
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Splits a string by newlines and adds a chosen number of tabs to each line.
}
\section{Creation notes}{
 First created in 2020-Oct while for building dynamic
SQL statements.
}

\examples{
\dontrun{
library(glue)
library(stringr)

tableRows <- glue::glue("
<tr class='rowSeperator'>
    <td>Test</td>
</tr>
")

glueIndentLines(tableRows, 2)
}
}
\author{
Jacob Peterson
}
