% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupby_rank.R
\name{groupby_rank}
\alias{groupby_rank}
\title{Helper for group_by |> mutate(rankID) |> filter(rankID)}
\usage{
groupby_rank(
  .data,
  ...,
  rankby,
  rankDesc = TRUE,
  filterIDs = NULL,
  removeRankCol = TRUE,
  doUngroup = TRUE
)
}
\arguments{
\item{.data}{data.frame/tibble}

\item{...}{group_by fields passed to group_by}

\item{rankby}{what column to rank by as character}

\item{rankDesc}{descending rank? TRUE/FALSE}

\item{filterIDs}{what rankID's to filter out. Default NULL}

\item{removeRankCol}{Remove rank column? TRUE/FALSE. Only used if filtering}

\item{doUngroup}{Ungroup the data? TRUE/FALSE}
}
\value{
tibble
}
\description{
Helper for group_by |> mutate(rankID) |> filter(rankID)
}
\section{Creation notes}{
 First created in 2020-12-11 while working in the
MARC-KC/CovidDataEntry Repository
}

\examples{
\dontrun{
groupby_rank(iris, Species, rankby = "Sepal.Length", rankDesc = TRUE, filterIDs = 1)
}
}
