% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renvHelpers.R
\name{is_renv_used}
\alias{is_renv_used}
\title{Is {renv} used in current directory/project?}
\usage{
is_renv_used(path = here::here())
}
\arguments{
\item{path}{Path to check. Defaults to result of \code{here::here()}}
}
\value{
Returns a TRUE/FALSE
}
\description{
Check if your current workspace is in a project with renv
enabled. This is kind of a hack and just looks for a renv folder, but it
should work.
}
\examples{
\dontrun{
is_renv_used()
}

}
