% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_date.R
\name{map_date}
\alias{map_date}
\alias{map2_date}
\title{map_date and map2_date}
\usage{
map_date(.x, .f, ...)

map2_date(.x, .y, .f, ...)
}
\arguments{
\item{.x, .y}{list or atomic vector.}

\item{.f}{A function, formula, or vector (not necessarily atomic).

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function. There
are three ways to refer to the arguments:
\itemize{
\item For a single argument function, use \code{.}
\item For a two argument function, use \code{.x} and \code{.y}
\item For more arguments, use \code{..1}, \code{..2}, \code{..3} etc
}

This syntax allows you to create very compact anonymous
functions. Note that formula functions conceptually take dots
(that's why you can use \code{..1} etc). They silently ignore
additional arguments that are not used in the formula expression.

If \strong{character vector}, \strong{numeric vector}, or \strong{list}, it is
converted to an extractor function. Character vectors index by
name and numeric vectors index by position; use a list to index
by position and name at different levels. If a component is not
present, the value of \code{.default} will be returned.}

\item{...}{Additional arguments passed on to methods.}
}
\value{
Returns a vector of dates
}
\description{
Wrappers for purrr::map and purrr::map2 to output dates
}
\examples{
\dontrun{
#Need to add, but very similar to any of the other purrr::map interfaces
}
}
