% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARCDataAPI.R
\name{MARCDataAPI_read}
\alias{MARCDataAPI_read}
\title{Download Data from MARCDataAPI}
\usage{
MARCDataAPI_read(url, query = NULL)
}
\arguments{
\item{url}{A url to a MARCDataAPI.}

\item{query}{A list of key-values for queries to append to the GET request URL. Currently untested.}
}
\value{
OBDC connection object returned from DBI::dbConnect()
}
\description{
Creates a GET request for a MARCDataAPI URL and returns the result as a tibble
}
\details{
See https://rpubs.com/ankc/480665 for how query should work theoretically.
}
\section{Creation notes}{
 First created in 2020-12-30 for easily accessing
data distributed by the MARCData API.
}

\examples{
\dontrun{
df <- MARCDataAPI_read('https://gis2.marc2.org/MARCDataAPI/api/covidcasedeathtest') \%>\% 
  dplyr::mutate(Date = as.Date(Date),
                LastUpdated = lubridate::as_datetime(LastUpdated),
                LastUpdated = lubridate::with_tz(LastUpdated, "America/Chicago")) 
df
}
}
\author{
Jacob Peterson
}
