% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcR-package.R
\docType{package}
\name{marcR-package}
\alias{marcR}
\alias{marcR-package}
\title{marcR: Helpful Functions Used By The MARC-GIS Team}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jacob Peterson \email{marc_gis@marc.org}

Authors:
\itemize{
  \item MARC-GIS \email{marc_gis@marc.org}
}

}
\keyword{internal}
