% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeLists.R
\name{mergeList}
\alias{mergeList}
\title{Merge (concatenate) lists elementWise}
\usage{
mergeList(x, y)
}
\arguments{
\item{x, y}{Lists with the same structure and element classes that can be
concatenated with \code{c}.}
}
\value{
Merged list with the same structure as x and y.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Wrapper around mapply to do element-wise concatenation of a list
}
\section{Creation notes}{
 First created on 2021-01-11 while working in the
BuildingFootprints DataDevelopment project.
}

\examples{
\dontrun{
xyz <- tibble::tribble(
    ~x,            ~y,           ~z,           ~xyz,
    11L,           integer(0),   integer(0),   11L,
    120L,          220L,         320L,         c(120L, 220L, 320L),
    c(13L, 14L),   integer(0),   33L,          c(13L, 14L, 33L),
    integer(0),    integer(0),   integer(0),   integer(0),
    15L,           integer(0),   integer(0),   15L,
    
    integer(0),    21L,          integer(0),   21L,
    integer(0),    22L,          integer(0),   22L,
    integer(0),    23L,          integer(0),   23L,
    integer(0),    24L,          integer(0),   24L,
    integer(0),    25L,          integer(0),   25L,
    
    integer(0),    integer(0),   31L,          31L,
    integer(0),    integer(0),   32L,          32L,
    integer(0),    integer(0),   33L,          33L,
    integer(0),    integer(0),   34L,          34L,
    integer(0),    integer(0),   35L,          35L
    )

mergeList(xyz[['x']],xyz[['y']])
mergeList(xyz[['x']],xyz[['z']])
Reduce(mergeList,xyz[c('x', 'y', 'z')])
identical(Reduce(mergeList,xyz[c('x', 'y', 'z')]), xyz[['xyz']])
}
}
