% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSMhelpers.R
\name{osm_keyValueDF}
\alias{osm_keyValueDF}
\title{Retrieve all Key-Value Pairs Open Street Map Data}
\usage{
osm_keyValueDF()
}
\value{
A data frame with two columns, 'key' and 'value'.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Downloads and creates a dataframe for all key-value pairs in OSM
data. This function is parameterless.
}
\section{Creation notes}{
 First created in 2020-11-03 with the script
OSMdownloadAndSFtoSQLInterface.R in the MARC-KC/HelpLibrary Repository
}

\examples{
\dontrun{
library(purrr)
library(tibble)
library(osmdata)

df <- osm_keyValueDF()
df
}
}
\author{
Jacob Peterson
}
