% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_timestamp.R
\name{random_timestamp}
\alias{random_timestamp}
\title{Random timestamp}
\usage{
random_timestamp(n, start_date = "2016-01-01", day_range = 365)
}
\arguments{
\item{n}{number of random timestamps}

\item{start_date}{a character string of length 1 detailing the start date in the format YYYY-MM-DD}

\item{day_range}{number of days after start_date to generate}
}
\description{
generate a random string timestamp of the format: 2016-02-29T22:53:28.835
}
\examples{

random_timestamp(5)
}
