% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biotyper_out.R
\name{read_biotyper_out}
\alias{read_biotyper_out}
\title{Read biotyper logfiles}
\usage{
read_biotyper_out(file)
}
\arguments{
\item{file}{path to file <chr>}
}
\description{
Bruker biotyper generates logs in `D:/Data/mbt-out/` if the folder exists.
These files can be used to survey answers from the machine.
}
\examples{

one_file <- read_biotyper_out("data-raw/biotyper_1.astm")

more_files <- list.files("data-raw",
  pattern = ".astm|.out", full.names = TRUE) \%>\%
  purrr::map_dfr(read_biotyper_out, .id = "run_id")
}
