% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cobas_lgd.R
\name{read_cobas_lgd}
\alias{read_cobas_lgd}
\title{read_cobas_lgd}
\usage{
read_cobas_lgd(file)
}
\arguments{
\item{file}{path to file <chr>}
}
\description{
Read lgd files from a COBAS Liat.
}
\examples{

curve_data <- read_cobas_lgd("data-raw/inf_b.lgd")
}
