% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cobas_rst.R
\name{read_cobas_rst}
\alias{read_cobas_rst}
\title{Title read_cobas_rst}
\usage{
read_cobas_rst(file)
}
\arguments{
\item{file}{path to file <chr>}
}
\description{
Reads a rst file from a COBAS Liat. Most information is available through
middleware, but information on invalid runs is currently not.
}
\examples{

result <- read_cobas_rst("data-raw/valid_pos.rst")

many_results <- list.files("data-folder",
  full.names = TRUE, pattern = ".rst"
) \%>\%
  purrr::map_dfr(read_cobas_rst)
}
