% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameByPairs.R
\name{renameByPairs}
\alias{renameByPairs}
\title{Changes column names based on vectors containing old and new colNames}
\usage{
renameByPairs(df, currentNames, newNames)
}
\arguments{
\item{df}{data.frame/tibble}

\item{currentNames}{Character vector of current column names}

\item{newNames}{Character vector of new column names}
}
\value{
A tibble
}
\description{
Changes column names based on vectors containing old and new colNames
}
\section{Creation notes}{
 First created in 2020-12-11 while working in the
MARC-KC/CovidDataEntry Repository
}

\examples{
\dontrun{
names(iris)
names(renameByPairs(iris, 
                    currentNames = c("Petal.Length", "Petal.Width"),
                    newNames = c("pl", "pw")))
}


}
