% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seleniumHelpers.R
\name{seleniumStart}
\alias{seleniumStart}
\alias{seleniumStop}
\title{Start/Stop Selenium Server}
\usage{
seleniumStart()

seleniumStop()
}
\description{
Provides a method to start and stop a selenium server. A lot of
this function is hardcoded and may not work in all environments. It will be
generalized have parameters added as it starts to get used more.
}
\details{
\describe{
\item{seleniumStart()}{Starts a firefox server on the
default port (4445). The function adds 2 objects to the global environment
(\code{rD} and \code{remDr}) that are used to manipulate and interact with
the server from your R code}
\item{seleniumStop()}{Removes the created objects from the R environment and cleans
up the detached server with command to kill any open java.exe tasks. This
is the most dangerouse part of this funciton and only works on a windows
computer}
}
}
