% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBIhelpers.R
\name{sf_readSQL}
\alias{sf_readSQL}
\title{Read Spatial SQL Table to sf}
\usage{
sf_readSQL(
  conn,
  schema,
  tableName,
  geomCol,
  whereClause = "",
  as_tibble = TRUE
)
}
\arguments{
\item{conn}{A \code{\link[DBI:DBIConnection-class]{DBIConnection}} object, as
returned by \code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{schema}{Name of schema that table is under. character vector of length
1.}

\item{tableName}{Name of table. character vector of length 1.}

\item{geomCol}{Character vector of length 1 containing the name of the
geometry column in the SQL table}

\item{whereClause}{The WHERE clause for the SQL Query}

\item{as_tibble}{Passed to \code{sf::st_read()}. Default is TRUE}
}
\value{
An sf object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Is basically a wrapper around
\code{\link[sf:st_read]{sf::st_read()}} that does a necessary
transformation on the SQL geometry datatype to read in as a sf table
}
\section{Creation notes}{
 First created in 2020-11-03 with the script
OSMdownloadAndSFtoSQLInterface.R in the MARC-KC/HelpLibrary Repository
}

\examples{
\dontrun{
library(DBI)

con <- connectODBC("<servername>.<databaseName>.<schemaName>")

#This is not set up to be a real example
tableColNames <- sf_readSQL(con, "<schemaName>", "<tableName>", "geom")
}
}
\author{
Jacob Peterson
}
