% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLhelpers.R
\name{SQL_createTable}
\alias{SQL_createTable}
\title{SQL_createTable}
\usage{
SQL_createTable(
  df = NULL,
  types = NULL,
  charLength,
  colNames = NULL,
  tableName,
  primaryKey = NULL
)
}
\arguments{
\item{df}{data.frame/tibble. Optional. Must have df or types specified.}

\item{types}{Character vector of new column types (readr-like). Optional.
Must have df or types specified.}

\item{charLength}{Vector of lengths used for character (NVARCHAR) types. If
using the df argument, it will also accept a named vector of lengths for
each column name with a character type.}

\item{colNames}{Character vector of columns names to include in database
table. Ignored if using the df argument.}

\item{tableName}{Name of database table you want to create.}

\item{primaryKey}{Name of the primary key for the table. Default Null does
not create one.}
}
\value{
A string containing the SQL code to create the specified table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Creates raw SQL query to be used to create a table
}
\section{Creation notes}{
 First created in 2020-12-15 while working in the
MARC-KC/CovidDataEntry Repository
}

\examples{
\dontrun{
head(iris)
irisHelper <- tibble::tribble(
  ~name,              ~length,  ~type,
  'Sepal.Length',     NA,       'd',
  'Sepal.Width',      NA,       'd',
  'Petal.Length',     NA,       'd',
  'Petal.Width',      NA,       'd',
  'Species',          '10',     'c'
)

SQL_createTable(df = iris, charLength = c("Species" = 10),
                tableName = "schema.iris", primaryKey = 'UniqueID')
SQL_createTable(types = irisHelper[['type']] , colNames = irisHelper[['name']],
                charLength = irisHelper[['length']], tableName = "schema.iris")
}

}
