% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLhelpers.R
\name{SQL_newColumnFormatter}
\alias{SQL_newColumnFormatter}
\title{SQL_newColumnFormatter}
\usage{
SQL_newColumnFormatter(df = NULL, types = NULL, names, charLength)
}
\arguments{
\item{df}{data.frame/tibble. Optional. Must have df or types specified}

\item{types}{Character vector of new column types (readr-like). Optional. Must have df or types specified}

\item{names}{Character vector of columns names to include in database table}

\item{charLength}{Vector of lengths used for character (NVARCHAR) types.}
}
\value{
A string containing partial SQL query for column definitions.
}
\description{
Creates the column definition part of a SQL query. Used as a helper for \code{SQL_createTable} and \code{SQL_addColumns}
}
