% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumNA.R
\name{sumNA}
\alias{sumNA}
\title{sumNA}
\usage{
sumNA(vec)
}
\arguments{
\item{vec}{numeric, double or integer vector}
}
\value{
Sum of input vector
}
\description{
Wrapper around sum that defaults na.rm to TRUE
}
\section{Creation notes}{
 First created in 2020-11 while working in the
MARC-KC/COVID-10_MARC Repository
}

\examples{
\dontrun{
sumNA(iris[['Sepal.Length']])
sumNA(c(NA, TRUE, FALSE, TRUE))
}
}
