% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeColTypes.R
\name{changeColType}
\alias{changeColType}
\title{Change Column Types}
\usage{
changeColType(df, cols, types)
}
\arguments{
\item{df}{data.frame/tibble}

\item{cols}{Character vector of columns in df to change}

\item{types}{Character vector of new column types (readr-like)}
}
\value{
A data.frame/tibble
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Changes column types based on vectors of column names and readr types
}
\section{Creation notes}{
 First created in 2020-12-11 while working in the
MARC-KC/CovidDataEntry Repository
}

\examples{
\dontrun{
class(iris[['Species']])

irisNew <- changeColType(iris, cols = c("Species"), types=c('c'))
class(irisNew[['Species']])
}

}
