% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBIhelpers.R
\name{DBI_AppendSFtoTable}
\alias{DBI_AppendSFtoTable}
\title{Append sf table to a Microsoft SQL table}
\usage{
DBI_AppendSFtoTable(
  conn,
  sfTable,
  schema,
  tableName,
  createTableQuery = NULL,
  warnings = getOption("warn")
)
}
\arguments{
\item{conn}{A \code{\link[DBI:DBIConnection-class]{DBIConnection}} object, as returned by \code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{sfTable}{sf object to append to specified database table}

\item{schema}{Name of schema that table is under. character vector of length
1.}

\item{tableName}{Name of table. character vector of length 1.}

\item{createTableQuery}{Currently unused. Always NULL}

\item{warnings}{Warning key. Use \code{-1} to hide warning messages in function.}
}
\description{
Append sf table to a Microsoft SQL table
}
\section{Creation notes}{
 First created on 2020-11-03 with the script
OSMdownloadAndSFtoSQLInterface.R in the MARC-KC/HelpLibrary Repository
}

\examples{
\dontrun{
library(DBI)

con <- connectODBC("DB_<databaseName>.<schemaName>")

#This is not set up to be a real example
tableColNames <- DBI_AppendSFtoTable(con, sfTable, "<schemaName>", "<tableName>")
}
}
\author{
Jacob Peterson
}
