% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBIhelpers.R
\name{DBI_getOBDCtable}
\alias{DBI_getOBDCtable}
\title{Get Table from OBDC connection}
\usage{
DBI_getOBDCtable(conn, query, roundRealDigits = NULL)
}
\arguments{
\item{conn}{A \code{\link[DBI:DBIConnection-class]{DBIConnection}} object, as
returned by \code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{query}{SQL Query to server to request table}

\item{roundRealDigits}{Optionally round digits of all doubles. Takes an integer to fill in the \code{digits}}
}
\value{
Tibble of a database table
}
\description{
Get Table from OBDC connection
}
\section{Creation notes}{
 First created in 2020-12-11 while working in the
MARC-KC/CovidDataEntry Repository
}

\examples{
\dontrun{ 

library(DBI)

con <- connectODBC("DB_<databaseName>.<schemaName>")

table <- DBI_getOBDCtable(con, "SELECT * FROM <schemaName>.<tableName>") 
}

}
\author{
Jacob Peterson
}
