% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutateCalcString.R
\name{mutateCalcString}
\alias{mutateCalcString}
\title{mutateCalcString}
\usage{
mutateCalcString(df, mutateName, mutateCalc)
}
\arguments{
\item{df}{data.frame/tibble to change}

\item{mutateName}{name of the column you want to change or create with the mutation}

\item{mutateCalc}{calculation (as a string) to be used for the mutation}
}
\value{
data.frame/tibble with mutated columns
}
\description{
Allows a mutation given a string for the calculation side of the mutation
}
\section{Creation notes}{
 First created in 2020-11 while working in the
MARC-KC/COVID-10_MARC Repository
}

\examples{
\dontrun{
mutateCalcString(iris, "sumLength", "Sepal.Length + Petal.Length")
}
}
