% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLhelpers.R
\name{SQL_createTable}
\alias{SQL_createTable}
\title{SQL_createTable}
\usage{
SQL_createTable(
  df = NULL,
  types = NULL,
  names,
  charLength,
  tableName,
  primaryKey = NULL
)
}
\arguments{
\item{df}{data.frame/tibble. Optional. Must have df or types specified}

\item{types}{Character vector of new column types (readr-like). Optional.
Must have df or types specified}

\item{names}{Character vector of columns names to include in database table}

\item{charLength}{Vector of lengths used for character (NVARCHAR) types.}

\item{tableName}{Name of database table you want to create.}

\item{primaryKey}{Name of the primary key for the table. Default Null does
not create one.}
}
\value{
A string containing the SQL code to create the specified table
}
\description{
Creates raw SQL query to be used to create a table
}
\section{Creation notes}{
 First created in 2020-12-15 while working in the
MARC-KC/CovidDataEntry Repository
}

\examples{
\dontrun{
head(iris)
irisHelper <- tibble::tribble(
  ~name,              ~length,  ~type,
  'Sepal.Length',     NA,       'd',
  'Sepal.Width',      NA,       'd',
  'Petal.Length',     NA,       'd',
  'Petal.Width',      NA,       'd',
  'Species',          '10',     'c'
)

SQL_createTable(df = iris, names = irisHelper[['name']],
   charLength = irisHelper[['length']],
   tableName = "schema.iris", primaryKey = 'UniqueID')
SQL_createTable(types = irisHelper[['type']] , names = irisHelper[['name']],
   charLength = irisHelper[['length']], tableName = "schema.iris")
}

}
